+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [IMPROVEMENT] Reduce number of view updates by filtering events from payload. See [#1328][]
+
 # 1.21.0 / 27-07-2023
 - [BUGFIX] Fix TracingUUID string format. See [#1311][] (Thanks [@changm4n][])
 - [BUGFIX] Rename _Datadog_Private to DatadogPrivate. See [#1331] (Thanks [@alexfanatics][])
@@ -472,6 +474,7 @@
 [#1311]: https://github.com/DataDog/dd-sdk-ios/pull/1311
 [#1315]: https://github.com/DataDog/dd-sdk-ios/pull/1315
 [#1331]: https://github.com/DataDog/dd-sdk-ios/pull/1331
+[#1328]: https://github.com/DataDog/dd-sdk-ios/pull/1328
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu